+++ projectEuler/problems/CProblem1.h
@@ -1,13 +1,15 @@
 #pragma once
 #include "CProblem.h"
 
+#include <cstdint>
+
 
 namespace euler
 {
 	class CProblem1 : public CProblem
 	{
 	public:
-		CProblem1(int  range);
+		CProblem1(int64_t  range);
 
 	private:
 		void calculate() override;
