+++ projectEuler/problems/CProblem3.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "CProblem.h"
 
+#include <cstdint>
 #include <set>
 
 
@@ -9,7 +10,7 @@ namespace euler
 	class CProblem3 : public CProblem
 	{
 	public:
-		CProblem3(size_t  range);
+		CProblem3(int64_t  range);
 
 	private:
 		void calculate() override;
