+++ projectEuler/problems/CProblem69.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "CProblem.h"
 
+#include <cstdint>
 #include <set>
 
 
@@ -9,7 +10,7 @@ namespace euler
 	class CProblem69 : public CProblem
 	{
 	public:
-		CProblem69(int  range);
+		CProblem69(int64_t  range);
 
 	private:
 		void calculate() override;
